move string_model fixture to api_clients/conftest
