+++ tests/conftest.py
@@ -1,12 +1,9 @@
 import os
 from datetime import datetime, timezone
-from unittest.mock import patch
 
 import hypothesis
 import pytest
 from hypothesis import strategies as st
-
-from enlyze.api_clients.base import ApiBaseModel
 
 hypothesis.settings.register_profile("ci", deadline=None)
 hypothesis.settings.load_profile(os.getenv("HYPOTHESIS_PROFILE", "default"))
@@ -26,12 +23,3 @@ datetime_before_today_strategy = st.datetimes(
 @pytest.fixture
 def auth_token():
     return "some-token"
-
-
-@pytest.fixture
-def string_model():
-    with patch(
-        "enlyze.api_clients.base.ApiBaseModel.parse_obj",
-        side_effect=lambda o: str(o),
-    ):
-        yield ApiBaseModel
